run: Gracefully handle malformed output from UART
Previously, if an invalid unicode character was received on the UART,
Python would raise an exception while trying to decode it to text.
Rather than doing that, use the 'replace' policy, where bad characters
are replaced with a standin.

This issue was seen while running kselftests::kvm::get-reg-list, which
outputs a bad character.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>