build: Fix dependency graph resolution for artifacts in build section
Previously, ${artifact:} macros were only being looked for in the
`params` section of a component. But the documentation says that they
are also supported in prebuild/build/postbuild/clean. So if an artifact
was directly consumed in one of those sections it would not be
considered when building the dependency graph, and therefore there would
be a race of production vs consumption of the artifact.

Fixes: 79e7a2d ("process, utils: Implement the "resolveb" action for the process command.")

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>