config: Add ability to build kselftests as part of linux-base.yaml
By adding the following as a layer above linux-base.yaml:

build:
  linux:
    prebuild:
      - export BUILD_KSELFTESTS=true

kselftests will now be built, packaged as a tgz and made available in
the package and with the ${artifact:KSELFTESTS} macro.

Note that not all tests currently build cleanly; there are some
x86-specific tests that (obviously) complain, and a few that complain
about missing headers (e.g. sys/capability.h). As a result the output is
rather noisy. But I think this is par for the course with kselftests. At
least all the commonly used tests compile happily.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>