utils: Use unique id for proc streams rather than friendly name
We are about to make the friendly name optional to avoid the need to
prepend the tag on each line of output. So in preparation, assign a
unique id to each process stream and use that to determine if we have
stream interleving mid-line. Since multiple streams could soon choose an
empty friendly name, we can no longer use that as a unique id.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>