config: Configure color coding on a per-terminal basis
Applying a color coding to each terminal's output can interfere with
coloring/escape sequences that an interactive shell running within the
FVP may want to apply. Although we already have the --no-color command
line option, this is not fine-grained enough, so allow each terminal to
specify a no_color property.

We then set this property for all terminals within the standard set of
configs that are stdinout.

The result is that the main interactive terminal is now the default text
color, other terminals are still colored differently, and any (e.g.)
`ls` output is now colored correctly.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>