run: Stop removing escape sequences by default
Our attempt to remove escape sequences from the terminal output streams
was motivated by wanting to stop EDK2 from clearing over previous output
(and generally leaving the terminal in a bad state for future output),
as well as not wanting the escape sequences to conflict with
shrinkwrap's use of color coding for different output sources.

But it was very broken, and meant that legitimate escape sequence
generators (e.g.) bash, ls, etc were being punished and the UX was
diminished.

Make it opt-in to remove escape sequences on a per-terminal basis. This
is specified in the config, using the new terminal::<name>::no_escapes
property. By default this is disabled. But for EDK2 based configs, we
set it to be enabled during boot until handoff to the kernel, which will
disable it.

Its still not perfect; if an escape sequence gets split across read
buffers, we will not notice it. But given we are attempting to strip
escape sequences far less often than we used to (only for configs with
EDK2 and only until EDK2 hands off to the kernel), perhaps no body will
notice.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>