config: Use dt-base for device tree with CCA configs
We were previously limited to using the TFA-bundled device tree when
using the CCA configs, because there is no way to "preload" the DT for
Linux due to having to use full TFA FW and not using EDK2. So the DT
comes from the FIP, and is used by TFA during its boot. This makes it
difficult to modify the kernel command line.

dt-base was missing a few node/properties that TFA requires to boot, so
let's dynamically overlay those nodes/properties in dt-base.yaml, then
we can use it as the main DT in the FIP.

This enables cca-3world (and therefore cca-4world) to use dt-base.yaml
and all of it's conveniences (e.g. kernel command line, initrd
specification, etc) when running CCA configs, although we still have a
dependecy on the work for TF-A to communicate available memory size to
RMM dynamically. This is in the master and main branches but not
expected to be tagged until the next release cycle. To work around this
for now, we are always passing mem=1G on the command line (see comment
in cca-3world.yaml).

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>