config: Enable and use virtio-rng to speed up crng init
Linux needs a source of entrophy to init the crng. For arch revisions
v8.5 and newer, we have RNDR CPU instruction which can be used, but for
older revisions nothing is currently available, and Linux resorts to
trying to measure jitter, which takes a long time.

Fix this by using the virtio-rng device on the FVP. Older builds don't
have it, but we don't expect people to be using these old builds with
Shrinkwrap. So enable the device with an FVP parameter, mark it enabled
in the DT and compile Linux with virtio-driver support built in. With
this combo in place, Linux can use the device to init the crng.

This doesn't solve the EDK2/ACPI config becuase Linux doesn't know about
the device's presence.

This also doesn't solve the case where people bring their own kernel,
which doesn't have the driver compiled it. I have a patch on list to add
this to the defconfig.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>