test: refactor the test to supply list of tuples in the configs
Refactor the test so that CONFIGS specifies a list of tuples with the
second element being an optional dictionary of RTVARs. This helps to do
ACPI boot test with ns-edk2.ymal which is default DT boot.

We just need to have "acpi=false" in the command line for this test case.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>