config: Use EDK2 as host bootloader for CCA configs
We now have a branch for EDK2 that carves out the required RAM for RMM,
so migrate cca-3world.yaml (and cca-4world.yaml) to use EDK2. As a
result, the command line is now an rtvar. By default, we still boot
Linux with the DT, but ACPI can be used by adding "acpi=force" at
runtime:

--rtvar CMDLINE="mem=1G console=ttyAMA0 earlycon=pl011,0x1c090000 root=/dev/vda ip=dhcp acpi=force"

Its implemented by hoisting cca-3world.yaml on top of ns-edk2.yaml and
removing all the now-uneccessary repetition.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>