config: edk2-base.yaml: Split acpica into its own component
acpica is a completely separate component to edk2, so model it as such.
Now that we support artifacts that are directories, it is easy enough to
expose the built tools to edk2.

This change will prevent duplication when we shortly introduce a new
component for the edk2 cca guest fw, which also depends on acpica.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>