config: Build guest EDK2 FW for CCA configs
EDK2 CCA Guest FW is the component that bootstraps a confidential realm
VM. Instead of passing the kernel directly to the VMM (e.g. kvmtool),
the FW is passed along with an encrypted disk image that contains the
kernel and rootfs. This first step builds the FW, but the user must
subsequently prepare the disk image.

Although very similar to the EDK2 FW we already build for the FVP,
Shrinkwrap does not support multiple instances of a single component.
Therefore we model this as a new component, edk2-cca-guest.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>