config: Build guest disk image for CCA configs
Build a disk image that contains a fat16 'boot' partition, containing
the Linux kernel as the EFI bootloader, and an ext4 'root' partition,
with the contents of a supplied ext4 fs, or a blank ext4 fs.

In future, this can be expanded to encrypt the root partition and sign
the kernel, etc.

The user passes in the root partition using the GUEST_ROOTFS btvar.