build: Split into separate build and buildall commands
We are about to introduce build-time variables (btvars), which are a
per-config set of variables that the user can override on the command
line at build time. This means that it won't be possible to implicitly
build multiple configs at the same time, because each config will need
its own set of btvars.

Our solution is to split 'build' into 2 commands; 'build' will take a
single config and optionally btvars on the command line, and build that
single config. 'buildall' will take a yaml file that declares a set of
configs and the btvars for each config, and build them all concurrently.

Start that change by defining the new command, buildall, and modify
build to take a single config. Most of the code remains common and lives
in buildall; We just need to create a list of one config and everything
looks the same from then on. btvars are not yet supported with this
change.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>