config: Improve macro substitution
Dramatically improve the macro substitution logic ahead of adding btvars
support. It is now possible to have artifacts that refer to other
artifacts and everything gets resolved correctly, as long as there are
no circular dependencies. Once we add btvars, it will be possible to set
those btvars to artifacts macros and everything will continue to work.

The code is also significantly cleaned up to the point where it mostly
makes sense, although still plently of room for improvement.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>