config: Add btvars to yaml schema
btvars are like rtvars, except they are used during build-time. They can
be defined in the config (under the new 'buildex:btvars:' key), with an
optional default value. Users can optionally override them on the
command line when invoking a build. (or at least they will be able to
once we enable that part in a future commit). btvars can be used in the
config through the macro system. It is even possible to set a btvar to
an ${artifact:*} macro, and things will resolve correctly.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>