test: Rework to use new buildall command
Now that we have separated build and buildall, rework the tests to use
buildall by generating a temporary configs.yaml. This allows us to
continue to benefit from building multiple configs in parallel.

While we are at it, add a mechanism to easily specify config-specifc
btvars, and update the cca-3world.yaml test to specify the GUEST_ROOTFS
btvar.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>