docs: Auto-generate rst files for the standard configs
Add a Python script, genconfigsstore.py, which reads the output of
`shrinkwrap inspect` and generates a corresponding rst file for
insertion into the docs. This script must be run to regenerate the docs
whenever a config is modified or added.

Also include the initially generated set of files.

In the long term, it would be preferable to run this on the fly, during
documenation building, but its not that easy to do with ReadTheDocs
currently.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>