config: Upgrade to TF-A/Hafnium 2.9.0 and RMM 0.3.0
The new Hafnium and RMM releases have moved where the built artifacts
end up, so we also need to adjust the artifact locations.

RMM 0.3.0 now supports the second RAM bank in the FVP so we can drop the
restriction to only allow the first 1G.

And TF-A now supports SME2 so we can remove the workaround for that too.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>
Co-Developed-by: Anders Roxell <anders.roxell@linaro.org>