config: Fix kvmtool dynamic linkage with zlib
kvmtool's makefile will auto-detect various optional libraries at
build-time and link them in if available. There is no way to disable
this auto-detection. One of these libraries is zlib.

Commit 65e14f2 ("docker: Add library dependecies for kselftests") added
libz-dev:arm64 to the build environment for the "full" docker image,
which causes kvmtool to detect and link against zlib. But zlib wasn't in
the rootfs, so kvmtool would fail at runtime with:

  lkvm: error while loading shared libraries: libz.so.1: cannot open
  shared object file: No such file or directory

This only happens when using the "full" image to build. Fix this by
adding zlib to the buildroot config.

Fixes: 65e14f2 ("docker: Add library dependecies for kselftests")
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>