docker: Update images to use latest component versions
Base image is updated to Debian Bookworm (which means the GNU/Linux
toolchain is now 12.2), FVP is updated to 11.24_11 and the GNU/None
toolchain is updated to 13.2.

Bookworm no lonfer allows installing pip packages as system packages, so
we set up a venv at /pyvenv and add it to the path so all pip packages
end up there.

Bookworm no longer has the "netcat" package so switch to
"netcat-openbsd", which the Bullseye "netcat" package was an alias for
anyway.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>