test: Refactor test.py for fine-graned arch overlays
Refactor test.py so that the CONFIGS table contains a range of arches
for which the config should be tested. This allows the future addition
of configs that only work from a given starting arch version, for
example.

As part of this work move all the rtvars into the CONFIGS table so that
we no longer need to special case bootwrapper.yaml. This means it can be
built in parallel and speeds up the testing.

For now, cca configs remain special-cased, since they don't take an arch
overlay; they are fixed.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>