config: Add debug overlays for tfa and rmm
By default, the release builds are used for both tfa and rmm. In some
circumstances it is useful to use the debug builds though. Add some
overlays for convenience to do this switch. Users can now do e.g.:

  $ shrinkwrap build cca-3world.yaml -o debug/tfa.yaml -o debug/rmm.yaml

Note the new "debug" directory convention for overlays of this type
within the config store.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>