config: Speed up TFA and TFTF builds
In the past, build failures have been observed when using parallel make
on systems with large numbers of cores for both tfa and tftf. As a
result, make has been limited to 1 job for both components. But this is
slow and the TFA team advise that it is known to be solid when running
parallel make with 8 jobs. So there is likely a dependency bug that only
manifests when there are a huge number of jobs.

Work around this by allowing parallel make upto a maximum of 8 jobs.
This significantly improves build times.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>