config: Allow btvars to be used as rtvar default values
This allows a single parameter to be specified that can be used at both
build-time and run-time. A btvar is defined and set as normal. And a
corresponding rtvar is defined with its default value set to the btvar
macro. The user can still override the rtvar at runtime if they want.

Example:

  buildex:
    btvars:
      MYPROP:
        type: string
        value: null
  build:
    mycomponent:
      params:
        myprop: ${btvar:MYPROP}
  run:
    rtvars:
      MYPROP:
        type: string
        value: ${btvar:MYPROP}
    params:
      -C myprop: ${rtvar:MYPROP}

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>