config: Allow buildex params from overlays
Previously, build/clean/process were all spec'ed to merge only the build
and run sections from any overlays. I believe this was intended to allow
future additions to the config to not cause issues (e.g. if a
'configure' command and section was ever added, it doesn't make sense to
pull that into a build since configure has to have already happened).

However when 'buildex' was added as a back compat extension for the
build section, it was not treated as part of the build section and
erroneously filtered out. Let's fix that.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>