run: Fix broken user ack when using telnet terminal
If any terminal type is set to 'telnet', Shrinkwrap prints the required
telnet command that the user needs to connect, and waits for the user to
acknowledge by pressing enter. It uses the standard Python input()
function for this. However, the process manager has already redirected
stdin so input() never receives any input and never returns.

Fix this by explicitly requesting the process manager deactivates stdin
around the input() call.

Fixes: c99dc74 ("run: Overhaul user input handling.")
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>