build, clean, process, run: Allow json overlay on command line
The --overlay option is supported by most commands and allows passing a
yaml file to use as an overlay on top of the main config. Sometimes the
desired overlay is very small and it is onerous to create a file. So
let's allow a json-encoded overlay to be passed directly on the command
line. If the overlay string starts with a '{' it is interpreted as a
json object, otherwise it is treated as a yaml file name.

This allows (e.g.):

$ shrinkwrap run --overlay '{"run":{"params":{"-C bp.pl011_uart0.out_file":"out0.log"}}}' ...

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>