logger: Add free_data() hook
To use the logger, the caller must first allocate instance data using
Logger.alloc_data(). Previously all of this data was trivial so had no
hook to free it. However, we are about to add a file descriptor to it,
so we need a hook to free the file when it goes out of scope.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>