build: Always sync repos to prescribed revision
Previously, Shrinkwrap would only sync a component's repos to the
prescribed revision when the repo was first cloned. Any subsequent
invocation would leave the repos as they were; the user was free to
checkout a different revision or even dirty the working directory. This
behaviour was by design, and intended to enable the development use
case, where a developer is actively working on the codebase and wants to
do incremental builds.

However, this behaviour has led to a lot of user confusion. Most
critically, when upgrading Shrinkwrap and getting updated configs with
new specified versions, Shrinkwrap would not update the repo revision
but apply new build directives, which may not match. Chaos.

So let's change the behaviour. From now on, Shrinkwrap will explicitly
sync all repos to the specified version on all build invocations. In a
future commit, a per-component opt-out option will be added to re-enable
the developer use case.

We first try to checkout the specified revision. If that fails, we fetch
from the remote, then try checking out again. This means we avoid the
fetch, which is slow, in the common case. Checkout's --force ensures
that any dirty files are reverted. Finally, the submodules voodoo
ensures that if the repo has submodules, they are all correct too.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>