build: Split per-component artifact copy step
Previously there was a single, global artifact copy step at the end of a
build process, which copied all artifacts to the package. The step
dependend on all component build steps to be complete. This meant that
every component would display "waiting..." in the UI until all other
components were built, then all would display the "copying artifacts..."
message before all changing to "done".

Let's split the artifact copying step so that it's per-component and
therefore only needs to depend on the specific component. Then
components can run to completion and display "done" when they are done
rather than wait for the others.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>