build: Fix output to logfile
Log files are per-script fragment and are named based on the component
to which they are associated. Until commit 7be0dac ("build: Split
per-component artifact copy step"), there was at most, 1 script fragment
associated with each component, so this all worked as intended. But that
commit changed the "copy" script fragment to be per-component instead of
global. This meant that there were multiple log files with the same name
and the latter ones overwrote the former ones.

Fix this by introducing a per-component log file number, which is
appended to the log file name. It starts at 0 and increments for each
script fragment that is bound to a component.

Fixes: 7be0dac ("build: Split per-component artifact copy step")
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>