fix: TF-A ARM_ARCH options for v9.0, v9.1, v9.2
From TF-A v2.11 restore the ability to specify
ARM_ARCH_MAJOR/ARM_ARCH_MINOR from TF-A command line for the desired
architecture extension.
This was gated by the fix [1] now merged.
Note config files for arch extensions beyond v9.2 are not updated by
this change and deserve doing it at a later time.

[1] https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/25886

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>