test: Improve test list format and test naming
To make test.py more suitable for use in CI, let's do a better job of
standardising the test names using type, config and overlay. And for run
tests, let's add a tag to identify the run configuration, since we often
run the same build in different configs.

By reflecting this hierachy in the test list, we also fix a previous
problem where the same build configuraiton was passed to Shrinkwrap
multiple times, so this should speed up the test run.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>