test: Run multiple fvps in parallel
When running on a system with enough RAM and multiple CPUs, we can do
multiple FVP runs in parallel to save text execution time. This is
useful for the CI. A new option, `--fvpjobs` is added to specify the
maximum number of parallel instances. When 1 (the default), shrinkwrap
continues to output directly to the console as before. When >1, the
entire output of each run is captured then dumpped to stdout out
atomically.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>