docker: Enhance build.sh to support kaniko
When building images on CI, we need to use Kaniko, since docker doesn't
work when running inside docker. So let's enhance the script to allow
the user to specify the driver. While we are at it, also allow the user
to specify the registry, since that will be needed for CI too, to store
temporary images. Additionally, ensure the script runs under sh, since
bash is not available in the kaniko container image.

For some reason `uname -p` returns "unknown" on CI runners, so let's
switch to the more standard `uname -m` instead.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>