docker: Cleanup after each kaniko invocation
Previously, the base-full image contained the directory structure for
the FVP package, but all the files were missing. Experimentation
concluded that Kaniko was getting confused because it has already
untared the fvp package for the base-slim image. If I comment out
building that image, then base-full contains all the required files.

So let's work around that issue by providing `--cleanup` to Kaniko,
which will cause it to clean its self up at the end of each invocation
so as not to confuse future invocations.

With this in place, the base-slim image is now correct.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>