ci: Build documentation in readthedocs.com
Current strategy is that all official docs builds from main branch are
against the "latest" version in readthedocs. But we also want to build
docs in other branches in the CI to ensure that there are no build
errors.

In the latter case, readthedocs only discovers the new branches (and
therefore new "versions") when it does a git pull. So we must force it
to build a version it knows about (main) to force the git pull before
building the target branch. We ensure that the target branch remains
inactive so that the built documentation is thrown away - we only care
about the build result.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>