test: Fix stdout encoding
test.py captures stdout/err from the FVP and dumps it to the test's
stdout. It was previously encoding as utf-8 at source but in some cases
an exception meant that stdout was remaining binary. Let's always
capture it in binary then unconditionally encode as utf-8 when we write
it to stdout. That way we should sidestep the write() error if an
exception had occured.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>