config: Update buildroot config to use BR2_DEFCONFIG
This change is optional and changes nothing functionally. However, it
makes it more easier to understand the steps. Instead of coping the
default config to the builddir as .config and then running olddefconfig,
one could just use BR2_DEFCONFIG and point it to the buildroot default
config and run defconfig step.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>