test: Include basic OPTEE without FF-A into test vector
A configuration with OPTEE in secure EL1 with TF-A in secure EL3 without
FF-A and secure EL2/Hafnium is being adding. Let us add the same into
the test vector for build and boot test coverage.

There is no ACPI binding for this non-FFA configuration of running OPTEE
and hence no ACPI coverage needed for it.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>