docs: document explicitely the linux kernel/9p needed workaround.
The 9p driver in the linux kernel currently has an issue, where the guest
image is not shared properly with the realm guest EFI, preventing the
realm guest from booting. This patch push forward the workaround
(modifying the host rootfs) until a fix is implemented in the linux
kernel. The method of sharing a folder thru 9p is still left here
for reference though, as this will hopefully be fixed at some point.

Signed-off-by: Arnaud de Grandmaison <arnaud.degrandmaison@arm.com>