config: edk2: Remove EDK2 shell timeout to start kernel immediately
EDK2 PCD(Platform Configuration Data) `PcdShellDefaultDelay` controls
the default delay the shell will offer to the user at the start to check
if the user wishes to cancel the script autostart `startup.nsh`

The default value of 5 is set. However, when running FVP models with
shrinkwrap, there is no need for such a delay. Currently one needs to
enter any key to start the kernel immediately without that 5s delay.

Let us override the value by setting it to 0. If required, it can
changed to desired timeout using the overlays.

While at it, also override `PcdPlatformBootTimeOut` for the ArmVirtKvmTool
platform used in CCA 3-world config.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>