config: edk2: Build Uefi with Shell enabled as default boot option
EDK2 PCD(Platform Configuration Data) `PcdUefiShellDefaultBootEnable`
controls if the shell is enabled or disabled as default boot option.

By enabling the same, we can get rid of the pre-canned flash image
loaded into the model that allows to configure UEFI Shell as default
boot option.

All the changes to configstore/*.rst are auto-generated with
genconfigstore.py.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>