config: Split cca-3world.yaml to enable cca development use cases
Split cca-3world.yaml into 3 discrete layers:

  - tfa-rme.yaml
      Builds TF-A, TF-RMM, along with the required device tree,
      configured for RME with Armv9.2 architecture extension. Intended
      as a base config from which a variety of concrete cca configs can
      be built.
  - cca-edk2.yaml
      Augments tfa-rme.yaml by adding edk2 as the non-secure EL2
      bootloader. This is a concrete config that can be built and used
      to load a Linux kernel.
  - cca-3world.yaml
      Semantically exactly the same as previously; a full cca demo
      stack. But now its is composed using cca-edk2.yaml (and implicitly
      tfa-rme.yaml).

This approach raises an issue when including ns-edk2.yaml in
cca-edk2.yaml, because both it and tfa-rme.yaml include tfa-base.yaml.
We end up with 2 copies and it gets built twice, due to the merge rules
for lists. Solve this by splitting ns-edk2.yaml into a base part and a
concrete part. TF-A is provided by the concrete part only, so that the
base part can also be used by cca-edk2.yaml.

Co-developed-by: Sona Mathew <sonarebecca.mathew@arm.com>
Signed-off-by: Sona Mathew <sonarebecca.mathew@arm.com>
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>