configs: Drop dummy DEPENDS_ON_PARAM paramter from ffa-hafnium-optee
The dummy DEPENDS_ON_PARAM paramter is not used by TF-A but was added
just to ensure OPTEE is built before TF-A as it is used indirectly via
SP_LAYOUT_FILE. However, this is not needed as the prebuild section that
generates the SP_LAYOUT_FILE already has the dependency built.

This was added during the development before the prebuild section
gained the generation of the SP_LAYOUT_FILE instead was using the static
file.

Drop the dummy DEPENDS_ON_PARAM paramter from ffa-hafnium-optee as it is
no longer needed.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>