config: buildroot: Bump the version from 2022.05.3 to 2024.05.2
The current buildroot version is atleast 2 years old now. While most of
the things are just fine and continue to work with this version, the
recently added OPTEE tests has some indirect dependency.

Buildroot v2022.05.3 uses OPTEE tests v3.17.0 which fails running some
tests and they are fixed in later versions. Using this as the opportunity
bump the buildroot versions from 2022.05.3 to 2024.05.2.

As a result the OPTEE test versions are updated to v4.2.0.

Also regenerate the buildroot.config using savedefconfig while we are
updating the version. The glibc is the default library now and hence
the generated file drops the entry for the same.

Note the step for the regeneration is also added so that one can just
uncomment it when needed.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>