config: Drop the oldefconfig build step in buildroot
Initially, the build didn't use the saved/shorter form of defconfig to
set up the buildroot. Instead it copied a saved version of `.config`.
Commit 808f258ad04a ("config: Update buildroot config to use BR2_DEFCONFIG")
moved to using BR2_DEFCONFIG and running just defconfig step in prebuild
section.

However it left the olddefconfig build step which is okay as long as
no overlays add any additional config modifications in the prebuild
step. If there are any additional modifications in the prebuild step,
they will be overridden if olddefconfig is executed.

Drop the olddefconfig step to preserve any additional config changes in
the prebuild section.

This doesn't have any functional change if additional config changes are
not present in the prebuild section.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>