test: Add assets config file to generate all the test assets and builds
Currently `genassets.sh` script clones all the required repositories and
builds the assets required to run the test script. It works well but it
is easy to go out of sync with the linux-base and buildroot config files.
In order avoid that, each time those configs are updates, `genassets.sh`
also needs to be updated or we can start using shrinkwrap itself to build
these assets.

Add test-assets config file to generate all the test assets and builds
so that we can remove all the duplicated steps in `genassets.sh`.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>