build: Add 'sync' config option for each build component
By default all components are synchronized. Passing --no-sync=component
or --no-sync-all on the command-line disables syncing the specified
components. When using a lot of custom user source directories, passing
a selection of --no-sync parameters can be tedious. Add a 'sync' config
option which defaults to true, and can be overridden by a user config.
For example a local.yaml config that describes local repositories could
contain:

	build:
	  component_1:
	    sourcedir: /path/to/component_1
	    sync: false
	  component_2:
	    sourcedir: /path/to/component_2
	    sync: false

sync is a tristate: true, false, force. Since yaml automatically
translates 'true' and 'false' to python bools, internal representation
is a bit awkward but not problematic.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>