config: Update guest filesystem resizing steps after buildroot 2024.05 upgrade
The buildroot v2024.05 uses/builds newer versions of filesystem handling
tools. As a result, sometime when running the steps to resize the guest
filesystem from the copy of host filesystem, the listed steps(mainly
e2fsck and resize2fs) may fail due to incompatibilities.

Document the ways to fix the incompatibilities by using the tools from
the buildroot built copy of host tools which matches the versions correctly.

Reported-by: Olivier Deprez <olivier.deprez@arm.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>