feature: add an optional 'image' field to config files.
Some configurations require specific docker images to build (e.g.
cca-4worl.yaml needs shrinkwraptool/base-full), and this requirement is
exposed on the command line, as the user has to specify the image when
invoking shrinkwrap. This seems unnecessary and error prone.

This patch adds support for an 'image' keyword in the configuration file. If
no image is specified on the command line, shrinkwrap will use the image
from the recipe (or a default one if none is specified). When merging
multiple configurations, conflicts (different image names) are detected
and reported as we have no way of merging this requirement. At this time,
in case of conflict there is no way to resolve it by forcing the image to
use from the command line.

Signed-off-by: Arnaud de Grandmaison <arnaud.degrandmaison@arm.com>